# src/main/java/co/edu/uniquindio/proyectoprogramacion/dto/auth/LoginRequestDTO.java
package co.edu.uniquindio.proyectoprogramacion.dto.auth;

import jakarta.validation.constraints.NotBlank;
import lombok.Getter;
import lombok.Setter;

@Getter
@Setter
public class LoginRequestDTO {

    @NotBlank(message = "El username es obligatorio")
    private String username;

    @NotBlank(message = "La contraseña es obligatoria")
    private String password;
}