# src/main/java/co/edu/uniquindio/proyectoprogramacion/dto/common/ApiErrorDTO.java
package co.edu.uniquindio.proyectoprogramacion.dto.common;

import lombok.*;

import java.time.LocalDateTime;

@Getter
@Setter
@NoArgsConstructor
@AllArgsConstructor
@Builder
public class ApiErrorDTO {
    private LocalDateTime timestamp;
    private int status;
    private String error;
    private String message;
    private String path;
}