# src/main/resources/openapi.yml
openapi: 3.0.3
info:
  title: API Gestión de Solicitudes Académicas
  version: 2.1.0
  description: |
    API REST para sistema de triage y gestión de solicitudes académicas.

    Funcionalidades principales:
    - Registro de solicitudes académicas
    - Consulta y seguimiento del estado de solicitudes
    - Clasificación de solicitudes
    - Priorización de solicitudes
    - Asignación de responsables
    - Gestión controlada del ciclo de vida
    - Cierre formal de solicitudes
    - Historial auditable
    - Sugerencias asistidas por motor de reglas / IA opcional
    - Autenticación y autorización con JWT
servers:
  - url: http://localhost:8080
    description: Servidor local
...
security:
  - bearerAuth: []
paths:
  /api/auth/register:
    post:
      tags: [Autenticación]
      summary: Registrar nuevo usuario
      description: |
        Registra un nuevo usuario en el sistema.
        **Endpoint público**.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Usuario registrado correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseUsuarioResponse'
        '400':
          description: Datos inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: Conflicto por duplicidad
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /api/auth/login:
    post:
      tags: [Autenticación]
      summary: Iniciar sesión y obtener JWT
      description: |
        Autentica al usuario y retorna un token JWT.
        **Endpoint público**.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login exitoso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseLoginResponse'
        '400':
          description: Datos inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: Credenciales inválidas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /api/auth/me:
    get:
      tags: [Autenticación]
      summary: Obtener información del usuario autenticado
      description: |
        Retorna la información básica del usuario autenticado por JWT.
      responses:
        '200':
          description: Usuario autenticado obtenido correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseAuthMeResponse'
        '401':
          description: No autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    RolUsuario:
      type: string
      enum: [ESTUDIANTE, ADMINISTRATIVO, COORDINADOR]
      example: ESTUDIANTE
...
    ApiResponseBase:
      type: object
      required: [success, message, timestamp]
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Operación exitosa
        timestamp:
          type: string
          format: date-time
          example: "2026-03-14T15:30:00"
    ApiError:
      type: object
      required: [timestamp, status, error, message, path]
      properties:
        timestamp:
          type: string
          format: date-time
          example: "2026-03-14T15:30:00"
        status:
          type: integer
          example: 409
        error:
          type: string
          example: Conflict
        message:
          type: string
          example: La solicitud ya se encuentra cerrada y no puede modificarse
        path:
          type: string
          example: /api/solicitudes/1/cerrar
    ApiResponseLoginResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponseBase'
        - type: object
          required: [data]
          properties:
            data:
              $ref: '#/components/schemas/LoginResponse'
    ApiResponseAuthMeResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponseBase'
        - type: object
          required: [data]
          properties:
            data:
              $ref: '#/components/schemas/AuthMeResponse'
    ApiResponseUsuarioResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponseBase'
        - type: object
          required: [data]
          properties:
            data:
              $ref: '#/components/schemas/UsuarioResponse'
...
    RegisterRequest:
      type: object
      required:
        - username
        - password
        - nombreCompleto
        - identificacion
        - rol
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 50
          example: estudiante1
        password:
          type: string
          minLength: 4
          maxLength: 100
          example: "123456"
        nombreCompleto:
          type: string
          minLength: 5
          maxLength: 100
          example: Juan David Pérez
        identificacion:
          type: string
          minLength: 5
          maxLength: 20
          example: "1090123456"
        rol:
          $ref: '#/components/schemas/RolUsuario'
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 50
          example: admin1
        password:
          type: string
          minLength: 4
          maxLength: 100
          example: "123456"
    LoginResponse:
      type: object
      required: [token, type, username, roles]
      properties:
        token:
          type: string
          example: eyJhbGciOiJIUzI1NiJ9.token.signature
        type:
          type: string
          example: Bearer
        username:
          type: string
          example: admin1
        roles:
          type: array
          items:
            $ref: '#/components/schemas/RolUsuario'
          example: [ADMINISTRATIVO]
    AuthMeResponse:
      type: object
      required: [id, username, nombreCompleto, identificacion, authenticated, roles, activo]
      properties:
        id:
          type: integer
          format: int64
          example: 1
        username:
          type: string
          example: admin1
        nombreCompleto:
          type: string
          example: Administrador Principal
        identificacion:
          type: string
          example: "1090123456"
        authenticated:
          type: boolean
          example: true
        roles:
          type: array
          items:
            $ref: '#/components/schemas/RolUsuario'
          example: [ADMINISTRATIVO]
        activo:
          type: boolean
          example: true
    UsuarioResponse:
      type: object
      required: [id, username, nombreCompleto, identificacion, rol, activo]
      properties:
        id:
          type: integer
          format: int64
          example: 1
        username:
          type: string
          example: estudiante1
        nombreCompleto:
          type: string
          example: Juan David Pérez
        identificacion:
          type: string
          example: "1090123456"
        rol:
          $ref: '#/components/schemas/RolUsuario'
        activo:
          type: boolean
          example: true
...